added basic animation
The code for this is pretty terrible. I have setInterval in a useEffect
triggering a dummy state change, which sets off another useEffect with a
requestAnimationFrame. This will then recompute the transformation
matrix, which triggers another useEffect which will finally render the
scene.

Also for some reason all the shaders were being reloaded on every tick
which was tanking performance so I fixed that... by making another two
useEffects.

It's rough out here.